+++ src/components/layout/index.ts
@@ -1,4 +1,5 @@
 export * from './box';
 export * from './row';
+export * from './column';
 export * from './col';
 export * from './layout.types';